Now using a pure lookAt approach

+++ src/classes/Scene.h
@@ -13,14 +13,6 @@ public:
         RASTERISED,
         RAY_TRACED
     };
-//    enum LightingMode {
-//        DEFAULT,
-//        HARD_SHADOWS,
-//        PROXIMITY,
-//        ANGLE_OF_INCIDENCE,
-//        AMBIENT,
-//        SPECULAR
-//    };
     float width;
     float height;
     RenderMode renderMode;
@@ -29,6 +21,6 @@ public:
     Camera camera;
     DrawingWindow window;
     Scene(float width, float height, RenderMode renderMode, Light light, std::vector<ModelTriangle> triangles, Camera camera);
-    void moveLight(Camera::Axis axis, int sign);
+    void moveLight(Camera::Axis axis, float sign);
     void draw();
 };
